Move data preprocessing functions to seperate module
